Fixed issue with search task
We cannot use CurrentCulture when accessing the method from API without defining what language currentculture is,
a quickfix for this now is to just use DefaultLanguageID. But a proper way to fix this would be to actually pass
channel in the request header (litium-request-context - set in http.js on the client and read in RequestModelHandler on the server).
However the quickfix will do to keep this task as simple as possible and focused on building the query.